+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+## Version 1.7.0 (Oct 17, 2022 JST)
+
+- Add new constructor without specify in/out callbacks.
+- Add an 8bit in/out port method: `setPort8Callback`
+- Add an 16bit in/out port method: `setPort16Callback`
+
 ## Version 1.6.0 (Oct 16, 2022 JST)
 
 - Use `std::function` ... `constructor, addBreakPoint, addBreakOperand, setDebugMessage, setConsumeClockCallback, addReturnHandler, addCallHandler`